requirements.txt
# Project dependencies
praw>=7.7.0
python-dotenv>=1.0.0

# Dev
pytest
pytest-cov
